+++ Sandswept/Survivors/Ranger/States/Secondary/Release.cs
@@ -14,6 +14,7 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
         public bool hasFired = false;
         private GameObject tracerEffect;
         private GameObject impactEffect;
+        private GameObject muzzleFlash;
         private Transform modelTransform;
 
         public override void OnEnter()
@@ -45,26 +46,31 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
                     default:
                         tracerEffect = ReleaseVFX.tracerPrefabDefault;
                         impactEffect = ReleaseVFX.impactPrefabDefault;
+                        muzzleFlash = DirectCurrentVFX.muzzleFlashPrefabDefault;
                         break;
 
                     case "RANGER_SKIN_MAJOR_NAME":
                         tracerEffect = ReleaseVFX.tracerPrefabMajor;
                         impactEffect = ReleaseVFX.impactPrefabMajor;
+                        muzzleFlash = DirectCurrentVFX.muzzleFlashPrefabMajor;
                         break;
 
                     case "RANGER_SKIN_RENEGADE_NAME":
                         tracerEffect = ReleaseVFX.tracerPrefabRenegade;
                         impactEffect = ReleaseVFX.impactPrefabRenegade;
+                        muzzleFlash = DirectCurrentVFX.muzzleFlashPrefabRenegade;
                         break;
 
                     case "RANGER_SKIN_MILEZERO_NAME":
                         tracerEffect = ReleaseVFX.tracerPrefabMileZero;
                         impactEffect = ReleaseVFX.impactPrefabMileZero;
+                        muzzleFlash = DirectCurrentVFX.muzzleFlashPrefabMileZero;
                         break;
 
                     case "RANGER_SKIN_SANDSWEPT_NAME":
                         tracerEffect = ReleaseVFX.tracerPrefabSandswept;
                         impactEffect = ReleaseVFX.impactPrefabSandswept;
+                        muzzleFlash = DirectCurrentVFX.muzzleFlashPrefabSandswept;
                         break;
                 }
             }
@@ -172,12 +178,14 @@ namespace Sandswept.Survivors.Ranger.States.Secondary
                 attack.damageType.damageSource = DamageSource.Secondary;
                 attack.damageType.AddModdedDamageType(Electrician.Electrician.LIGHTNING);
 
-                AddRecoil(3f + 0.15f * buffCount, 3f + 0.15f * buffCount, 0f, 0f);
-
                 characterMotor?.ApplyForce((-4500f - 175f * buffCount) * aimDirection, false, false);
 
                 attack.Fire();
             }
+
+            EffectManager.SimpleMuzzleFlash(muzzleFlash, gameObject, "Muzzle", transmit: true);
+
+            AddRecoil(3f + 0.15f * buffCount, 3f + 0.15f * buffCount, 0f, 0f);
 
             characterBody.SetBuffCountSynced(Buffs.Charge.instance.BuffDef.buffIndex, Mathf.Max(0, buffCount - DirectCurrent.maxCharge));
         }