+++ Sandswept/Utils/Keywords.cs
@@ -30,11 +30,11 @@ namespace Sandswept.Utils
 
         public static void SetupKeywords()
         {
-            OverdriveFormPrimary.Add("<style=cKeywordName>Fire a rapid stream of bullets for $sd90% damage$se. $suFire rate and ignite chance increase with heat$se.".AutoFormat());
-            OverdriveFormSecondary.Add("<style=cKeywordName>$sdIgnite$se. $suReduce current heat by 50%$se. Fire a spread of heat for $sd4x300% damage$se. $suBurst count increases with heat spent$se.".AutoFormat());
-            OverdriveFormAltSecondary.Add("<style=cKeywordName>$sdIgnite$se. $suReduce current heat by 50%$se. Fire off a blazing orb for $sd1200%$se damage that $sdengulfs$se the ground on impact for $sd250%$se damage per second. $suDamage increases with heat spent$se.".AutoFormat());
-            OverdriveFormUtility.Add("<style=cKeywordName>$suAgile$se. $sdStunning$se. $suSidestep$se a short distance and deal $sd390% damage$se. Hitting enemies generates $rc5 Charge$ec and $sdextends$se the dash. $suDash distance increases with heat$se.".AutoFormat());
-            OverdriveFormSpecial.Add("<style=cKeywordName>$suAgile$se. $sdIgnite$se. $suConsume all heat and exit overdrive$se. Release a $sdfire nova$se around you that deals $sd300%$se damage. $suDamage increases with heat spent$se.".AutoFormat());
+            OverdriveFormPrimary.Add("<style=cKeywordName>Overdriven Form</style>Fire a rapid stream of bullets for $sd90% damage$se. $suFire rate and ignite chance increase with heat$se.".AutoFormat());
+            OverdriveFormSecondary.Add("<style=cKeywordName>Overdriven Form</style>$sdIgnite$se. $suReduce current heat by 50%$se. Fire a spread of heat for $sd4x300% damage$se. $suBurst count increases with heat spent$se.".AutoFormat());
+            OverdriveFormAltSecondary.Add("<style=cKeywordName>Overdriven Form</style>$sdIgnite$se. $suReduce current heat by 50%$se. Fire off a blazing orb for $sd1200%$se damage that $sdengulfs$se the ground on impact for $sd250%$se damage per second. $suDamage increases with heat spent$se.".AutoFormat());
+            OverdriveFormUtility.Add("<style=cKeywordName>Overdriven Form</style>$suAgile$se. $sdStunning$se. $suSidestep$se a short distance and deal $sd390% damage$se. Hitting enemies generates $rc5 Charge$ec and $sdextends$se the dash. $suDash distance increases with heat$se.".AutoFormat());
+            OverdriveFormSpecial.Add("<style=cKeywordName>Overdriven Form</style>$suAgile$se. $sdIgnite$se. $suConsume all heat and exit overdrive$se. Release a $sdfire nova$se around you that deals $sd300%$se damage. $suDamage increases with heat spent$se.".AutoFormat());
         }
     }
 }